Update project, batch file, example output to Unity 2017.1.0f3
The repro has changed a little; it now happens when adding a script,
but doesn't happen when deleting one. I've left the old repro in just in case,
as I have a feeling I'll still be reporting this bug several times a year for
the next decade.

Sorry, commit messages shouldn't be the place for sarcasm.